【feat】 项目重构，优化页面布局和数据库字段

+++ datai-scenes/datai-scene-salesforce/datai-salesforce-integration/src/main/java/com/datai/integration/controller/DataiIntegrationBatchController.java
@@ -27,7 +27,7 @@ import io.swagger.v3.oas.annotations.Operation;
  * 数据批次Controller
  * 
  * @author datai
- * @date 2025-12-22
+ * @date 2025-12-24
  */
 @RestController
 @RequestMapping("/integration/batch")
@@ -70,7 +70,7 @@ public class DataiIntegrationBatchController extends BaseController
     @Operation(summary = "获取数据批次详细信息")
     @PreAuthorize("@ss.hasPermi('integration:batch:query')")
     @GetMapping(value = "/{id}")
-    public AjaxResult getInfo(@PathVariable("id") Long id)
+    public AjaxResult getInfo(@PathVariable("id") Integer id)
     {
         return success(dataiIntegrationBatchService.selectDataiIntegrationBatchById(id));
     }
@@ -106,7 +106,7 @@ public class DataiIntegrationBatchController extends BaseController
     @PreAuthorize("@ss.hasPermi('integration:batch:remove')")
     @Log(title = "数据批次", businessType = BusinessType.DELETE)
     @DeleteMapping("/{ids}")
-    public AjaxResult remove(@PathVariable( name = "ids" ) Long[] ids) 
+    public AjaxResult remove(@PathVariable( name = "ids" ) Integer[] ids) 
     {
         return toAjax(dataiIntegrationBatchService.deleteDataiIntegrationBatchByIds(ids));
     }
