+++ datai-scenes/datai-scene-salesforce/datai-salesforce-integration/src/main/java/com/datai/integration/mapper/DataiIntegrationBatchMapper.java
@@ -7,7 +7,7 @@ import com.datai.integration.domain.DataiIntegrationBatch;
  * 数据批次Mapper接口
  * 
  * @author datai
- * @date 2025-12-22
+ * @date 2025-12-24
  */
 public interface DataiIntegrationBatchMapper 
 {
@@ -17,7 +17,7 @@ public interface DataiIntegrationBatchMapper
      * @param id 数据批次主键
      * @return 数据批次
      */
-    public DataiIntegrationBatch selectDataiIntegrationBatchById(Long id);
+    public DataiIntegrationBatch selectDataiIntegrationBatchById(Integer id);
 
     /**
      * 查询数据批次列表
@@ -49,7 +49,7 @@ public interface DataiIntegrationBatchMapper
      * @param id 数据批次主键
      * @return 结果
      */
-    public int deleteDataiIntegrationBatchById(Long id);
+    public int deleteDataiIntegrationBatchById(Integer id);
 
     /**
      * 批量删除数据批次
@@ -57,5 +57,5 @@ public interface DataiIntegrationBatchMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deleteDataiIntegrationBatchByIds(Long[] ids);
+    public int deleteDataiIntegrationBatchByIds(Integer[] ids);
 }
