+++ datai-scenes/datai-scene-salesforce/datai-salesforce-integration/src/main/java/com/datai/integration/service/IDataiIntegrationFilterLookupService.java
@@ -7,7 +7,7 @@ import com.datai.integration.domain.DataiIntegrationFilterLookup;
  * 字段过滤查找信息Service接口
  * 
  * @author datai
- * @date 2025-12-22
+ * @date 2025-12-24
  */
 public interface IDataiIntegrationFilterLookupService 
 {
@@ -17,7 +17,7 @@ public interface IDataiIntegrationFilterLookupService
      * @param id 字段过滤查找信息主键
      * @return 字段过滤查找信息
      */
-    public DataiIntegrationFilterLookup selectDataiIntegrationFilterLookupById(Long id);
+    public DataiIntegrationFilterLookup selectDataiIntegrationFilterLookupById(Integer id);
 
     /**
      * 查询字段过滤查找信息列表
@@ -49,7 +49,7 @@ public interface IDataiIntegrationFilterLookupService
      * @param ids 需要删除的字段过滤查找信息主键集合
      * @return 结果
      */
-    public int deleteDataiIntegrationFilterLookupByIds(Long[] ids);
+    public int deleteDataiIntegrationFilterLookupByIds(Integer[] ids);
 
     /**
      * 删除字段过滤查找信息信息
@@ -57,5 +57,5 @@ public interface IDataiIntegrationFilterLookupService
      * @param id 字段过滤查找信息主键
      * @return 结果
      */
-    public int deleteDataiIntegrationFilterLookupById(Long id);
+    public int deleteDataiIntegrationFilterLookupById(Integer id);
 }
